+++ terraform/modules/indent_runtime_aws_lambda/variables.tf
@@ -31,3 +31,8 @@ variable "artifact" {
     deps_key     = string
   })
 }
+
+variable "timeout" {
+  type    = number
+  default = 120
+}