+++ packages/stable/indent-integration-auto-approval-email/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.1-canary.4](https://github.com/indentapis/integrations/compare/@indent/integration-auto-approval-email@0.0.1-canary.3...@indent/integration-auto-approval-email@0.0.1-canary.4) (2022-06-24)
+
+**Note:** Version bump only for package @indent/integration-auto-approval-email
+
+
+
+
+
+## [0.0.1-canary.3](https://github.com/indentapis/integrations/compare/@indent/integration-auto-approval-email@0.0.1-canary.2...@indent/integration-auto-approval-email@0.0.1-canary.3) (2022-06-24)
+
+**Note:** Version bump only for package @indent/integration-auto-approval-email
+
+
+
+
+
 ## 0.0.1-canary.2 (2022-06-23)